+++ common/src/main/java/net/arsenal/item/Group.java
@@ -11,7 +11,7 @@ import net.arsenal.ArsenalMod;
 import java.util.function.Supplier;
 
 public class Group {
-    public static Identifier ID = Identifier.of(ArsenalMod.NAMESPACE, "generic");
+    public static Identifier ID = new Identifier(ArsenalMod.NAMESPACE, "generic");
     public static String translationKey = "itemGroup." + ID.getNamespace() + "." + ID.getPath();
     public static RegistryKey<ItemGroup> KEY = RegistryKey.of(Registries.ITEM_GROUP.getKey(), ID);
     public static ItemGroup GROUP;